separating infrastructure from module in deployment script for 2.5
