+++ utils/versions/development/latest.json
@@ -1,1 +1,1 @@
-{"version":"2.5.0","createdAt":1617284509,"modules":[{"address":"0x5616b71ddB5b35712543E897D97C9cF4178Df3A6","name":"ArgentModule"}],"fingerprint":"0xcce8561d"}
+{"version":"2.5.0","createdAt":1617376728,"modules":[{"address":"0x432547ef8dAa759969a2C98967e690D9A8033c23","name":"ArgentModule"}],"fingerprint":"0x893c5437"}
